+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Add `NestedGroups` cop for detecting excessive example group nesting. ([@backus][])
 * Add `MaxNesting` configuration option for `NestedGroups` cop. ([@backus][])
 * Add `ExpectActual` cop for detecting literal values within `expect(...)`. ([@backus][])
+* Add `MultipleExpectations` cop for detecting multiple `expect(...)` calls within one example. ([@backus][])
+* Add `Max` configuration option for `MultipleExpectations`. ([@backus][])
 
 ## 1.6.0 (2016-08-03)